fix component handling for ucpdav1 when --with-webdav=serf
Both the webdav=neon and webdav=serf cases were actually using
the same ucpdav1.component file. And eabe1f6dc64ce70b87347859cec5
changed component handling for neon, but not for serf.
So separate component handling for neon and for serf, and also
do the eabe1f6 change also for the serf case.

Change-Id: Id5907cf1c8fbb93817e81295f71ad87191fe5c16
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/114525
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>