tdf#141945 sw: layout: check master frame when moving fly forward
The problem is that in the finished layout the fly frames are
positioned on the first page but are in SwPageFrame::m_pSortedObjs
of the second page.

Don't use FindPageFrameOfAnchor() because that looks up the follow-frame
that contains the anchor position.  This was unintentional; the idea
was to get flys anchored in subsequent paragraphs out of the way.

This situation where it's on a follow-frame of the same paragraph is
more complicated and less obvious, so don't try to solve it now.

(regression from 320c7d0ce56775642e487a432a2ea4ef4fd4ee74)

Change-Id: Ic6809a43c467955368807ccc82a7873c29fbfc2e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/115100
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>