tdf#142348 Use transparent background for Android Viewer, too
After

    commit fa45426de8a123094175b7cbfbe222e82a4299ee
    Date:   Mon Apr 26 10:43:11 2021 +0900

        lok: add rendering and input for form controls in Calc

instead of properly rendering Calc documents, Android Viewer was
only showing a blank area instead of the actual doc content.
(In case of a Calc doc containing form controls, they were
visible on top of an otherwise white background.)

Using a transparent background by default in Android Viewer
as well, just as is done for all other scenarios already,
makes this work as expected again
(except for another issue introduced with

    commit 5e15ada7e580aa8e741bb22cc1d188b930b30689
    Date:   Fri Apr 9 13:59:43 2021 +0200

        tdf#95517 android: Rework app/doc lifecycle handling

that will be handled in a separate commit).

The comment already mentioned that it was unclear why
that should not work for Android Viewer and I did not see
any apparent problem when testing a bit, so just drop the
special handling for Android Viewer here.

Change-Id: I3005c5517ce1198af99b27af541fe59eaa896da2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/115749
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>