tdf#139915 DOCX import: fix anchored obj position with to-char and TEXT_LINE
There were multiple problems here:

- commit 7231effd17fe2efba6ecce06fe47478f22a7aa41 (tdf#97371 DOCX
import: fix text covered by shape, 2016-01-28) disabled to-char
anchoring for TEXT_LINE (e.g. "alignment top, relative to line") because
changing the anchor type of a TextBox could not be changed, but this has
been implemented in sw/ in the meantime, so it can be dropped

- Now that the anchor type is to-char, we can set the vertical relation
to TEXT_LINE, but Word's positive value is "below line", and ours mean
"towards the top of the page, from the bottom of the line", so we need
to drop the workaround of commit
9e5edcad31f1f8b855e539407e13011e45cf61ad (tdf#135153 DOCX import: avoid
line-of-text relation with to-para anchoring, 2021-01-18)

Once these are in place, we can fix the remaining problem by inverting
the vertical position of the shape, which instantly fixes the
overlapping textboxes in the bugdoc.

Change-Id: I895abb76d3bd3bbe3a84e5c27a77df722b96226a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/116182
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins