tdf#142404 sw compat layout: ignore blank size only on one-liners
If the paragraph is longer than one line,
then do not IgnoreBlanksAndTabsForLineHeightCalculation,
which was introduced in LO 4.0.

This is a preliminary step for fixing bug 142404.

I found a few pre-existing unit tests that triggered this,
but none were good examples to use as proofs.

ooxmlexport7: 77219 - not sure where - no visual difference.
ooxmlexport10: 92157 compatibilityMode15, but visual too little.
ooxmlexport11: 88496 - not sure where.
ooxmlexport13: 121374_sectionHF2.doc -tabOverMargin in header

NOTE: This patch could be seen as a BAD THING in the case where
tabOverMargin normally hides EVERYTHING. We don't handle
that situation yet, so a very long series of tabs could
take up several lines of space (which isn't done in Word).
Now with this change it could take even more space.

Well, the proper fix would be to not show any of those.
Perhaps we could just set the width of the portion to
zero if it sails past the end of the page?
Anyway, the point is that this it-isn't-right-anyway
situation should not block pushing this patch through.

Change-Id: Iaea8e0edf78c8fbe319aadbc6d62fc0bdd180814
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/116317
Tested-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Justin Luth <justin_luth@sil.org>