tdf#142404 DOCX c15: TabOverMargin no longer true in 2013+
When compatibilityMode is 15, TabOverMargin no longer
seems to apply. This is a dramatic visual change in docx
that I didn't find any documentation about,
but the visual change is obvious enough proof.

LibreOffice started saving DOCX as c15 mode in 7.0.

[P.S. related TabOverflow also seems to be false with c15,
but it acts differently than what LO's tabOverflow
code does. That was discussed in a different patch
and seems to be a dead end, so I'm ignoring that aspect.
Way too many complications and effects on LO native mode.]

Change-Id: I5a0a6d695d6825444cf6a362a81803f306e6c6e3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/116337
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>