do not use delayed layout in TitledDockingWindow
The layout is only performed in Paint(), so if something uses
geometry before that happens (may happen with tiled rendering),
then is not up to date. I don't see why delay it anyway, there
seems to be usually just one resize, and only during init I see
two of them, which can presumably be fixed if needed.

Change-Id: I7133f3c6be41bc6480f86d1053b455d42c40b9ba
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/116529
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>