NFC compat cleanup: no need to specify default TabOverflow
This really confused me because it lead me to think
that this was something done for MS compatibility.

Well, that is only true in an off-handed way.
LibreOffice itself was changed to work similarly
to MS Word. So there is nothing special about how DOC
or DOCX/RTF are handled.

Since the compat settings are not saved or loaded
into MS Formats (i.e. it just takes the default value),
and since on an ODT save it also will just save
with the proper default value, there is no need
to specify "TabOverflow = true" in non-ODT import filters.

Only ooxmlexport16 has a unit test that reacts if
tabOverflow is false. That one is mine and it indicates
that the document would be better if tabOverflow was off,
so there are no examples of how tabOverflow improves a doc.

Change-Id: I97c25154108bc1ca0fcd3dfcff66fea0ea2bca7e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/116741
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>