tdf#141805 sw: hide rtl gutter UI for ODF documents
ODF doesn't track RTLness of the gutter explicitly, it infers this
setting from the writing direction. This means that in case the UI is
used to set it explicitly to a custom value, it'll be lost after
export + import.

Fix the problem by hiding the widget in the ODF case, so it doesn't
confuse users; it's still visible for Word formats which store this
explicitly.

Change-Id: I1316dcf461429af72498957c9b3eebcbdac794bb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/116799
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>