tdf#142721 restore SwShellTableCursor if the orig selection was a single cell
where the previous attempt of

commit d0e2b2e0dadd2dea93e09452f8c3530a23ffb71e
    tdf#142165 restore a SwShellTableCursor if the orig selection described that
    so that we are operating on the same selection that existed when the
    dialog was originally launched

wasn't sufficient to restore a single cell selection.

we continue to need to avoid the assert of

commit a27e747b256980ea868537997752174f3ec2cfcd
    tdf#140977 drop possible table-cursor before setting the new one

and support the multi-selection of

commit d67340937895f25d5ad1a6e8c3164ea91ad253d1
    tdf#140257 duplicate entire PaM ring when making copy

and support not scrolling to a different location on changing a table
page break style of

commit ca2f2c87537432a9f027dbae5bcdaf3b2544e51f
    tdf#135916 just set the target table as selection

and keep making it possible to remove a page break on a table of

commit 85d25947c1e789a18f25feeb87629e3080be269b
    tdf#135636 the selection at dialog-launch time is lost by dialog-apply time

all of which is necessitated by

commit ca97acc1a35dffdf94e7cef7f1061f613f646107
    lokdialog: Convert the Table -> Properties... to async exec.

Though; since d0e2b2e0dadd2dea93e09452f8c3530a23ffb71e where we started
using rSh.GetTableCrs if IsTableMode() then in practice
a27e747b256980ea868537997752174f3ec2cfcd probably cannot arise.  The
scenario of rSh.IsTableMode() changing between launch of the dialog and
applying the result of the dialog is presumably maybe theoretically
possible in e.g. an online scenario, but not in the normal user case,
but handled here anyway.

Change-Id: I12f0b6bc7e0e2f5bad45a88007bf6fe2cd3d3b0b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/116933
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>