drop firebird test of initial experimental period fdb-only format
Initially when the firebird embedded database support was first
experimental we saved the fdb firebird file format.

A test was added by 6f03a35d34deb765c1091e7154bd38bf3f5718fe around then
(2013) to test loading firebird odbs

Then with

commit 01f0b4bc5ace59ff845a509177f9a324ca9eb198
Date:   Thu Jun 2 11:56:10 2016 +0200

    connectivity: firebird: use ODS12 test database for Firebird 3

if building with firebird 3 a replacement odb was created in firebird
3.X ODS12 format to use instead of the 2.X ODS11 file.

Those fdb file formats were endian specific and database version
specific. Which was a well known problem documented in tdf#72987 and a
blocker for leaving 'experimental' at that point in time.

So finally with

commit 4da5af6ba4d47fc4e5676a7a6709b51b14d7cfcf
Date:   Thu Aug 11 12:02:56 2016 +0200

    tdf#72987 GSoC Use Firebird backup format

the save was changed to save the backup fbk file format and load
restores the database from that fbk file format (2016) when available.

later in

commit dd236fd025f1b38c453068ecfdb2b49e42fe9e78
Date:   Thu Jan 5 12:58:26 2017 +0100

    require firebird 3.0 for build (don't allow 2.5)

the firebird 2.X (ODS11) test of loading the fdb format was
dropped leaving just the firebird 3.X (ODS12) fdb case because firebird
3.X cannot directly load the ODS11 format

Now with the appearance of firebird 4 the same problem arises that
firebird doesn't load older binary formats directly.

In the end all we established here is that an odb containing only a fdb
is endian and version specific and a bad idea wrt compatibility so we
shouldn't do that. But we knew that, so embedded firebird databases were
an experiemental curosity until 4da5af6ba4d47fc4e5676a7a6709b51b14d7cfcf
enabled use of the fbk format.

We're not gaining anything with the test of loading the binary only case
because we know that's an unsustainable route we abandoned.

Change-Id: I9f1715a817392de9f3f6e891966bbd838a9ce933
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/116981
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>