sw image anchor type: consider a style-based custom value
The anchor type is determined during insertion, it's stored as direct
formatting after that.

We used to have a user-profile default for the anchor type. This commit
allows customizing the Graphics or OLE styles: if they specify a custom
anchor type, then that is user instead of the user profile setting.

This allows creating templates where the default depends on the used
template and not on the user profile.

The UI for this was added in commit
567742810233e10556876fe2292cb70d993ec716 (sw image anchor type: add
style UI for this, 2021-06-10).

Change-Id: Id05342a5f38dc6267cdbe68b248dc50b87854ce2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/117040
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins