Resolves: tdf#140361 use the DisableEditHyperlink state as of menu launch time
at context menu popup time set if the EditHyperlink entry should be
disabled and use that state if queried about it if EditHyperlink is
dispatched from the menu. So ignoring where the mouse currently happens
to be when the menu was dismissed.

The dispatch is done async, if at all, so also trigger an async Query
with Invalidate so at least one Query is ensured to reset the stored
state

similar to tdf#137445 which was for impress/draw

Change-Id: I43a144f1ac0a4db89cc5ab0ebeeae744719f5958
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/117653
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>