cross-toolset: fix real cross help building
Regression from commit a0b8356ba0d78348cb8ae094d5efa20541512267
("configure: Improve help-options handling").

We now must forward the help and extension configure options, as
this is needed by the cross toolset to know, if it needs to build
the HelpIndexer and HelpLinker executables.

Also get rid of gb_Helper_optional_for_host. While it's in theory
more correct, it makes gb_Helper_optional more complex, so we now
rely on the fact, that the config_build.mk BUILD_TYPE is a subset
of the config_host.mk BUILD_TYPE.

This should fix the Windows Arm64 build, which currently seem to
be the only cross-build including help.

Change-Id: I8b90923c891b43cd2655b2578015c9bd579aae7d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/117758
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>