tdf#120629 ms formats: better exporting of hindiVowels etc.
The big problem was that hindiVowels was imported as
CHARS_NEPALI, but CHARS_NEPALI was not exported in DOCX.

Another easy win was to DOC export both LO forms of russianLower
and both forms of russianUpper - just like DOCX.

The page number field only allows a subset of formats.
LO doesn't have support for Hex or DollarText formats.
The Ordinal/Cardinal formats were an easy win
for export, even though they don't import.
[Import seems to be in ww8par5.cxx's GetNumTypeFromName(),
except that is being passed an empty string.]

Change-Id: I176a70e0e05967414337c17d8b4acc87e20ead73
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/117793
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>