fix failure in JunitTest_unotools_complex
Temp files created from OTempFileService are always opened with StreamMode::SHARE_DENYALL.
So normally you can't open them twice.

But the JunitTest_unotools_complex unit test used to work because it exploited a pessimisation in OTempFileService,
where that code would close the file when we read to the end.

Which meant that the unit test could open it again and read it.

However, in
    commit baf17b5564d0c826aa07a6eee8f7ce5d7674181f
    Date:   Sun Jun 20 18:51:12 2021 +0200
    tdf#135316 remove OTempFileService pessimisation
I removed that pessimisation.

So make the share mode a little more permissive.

Change-Id: I297a5c9c0505816b399fad29414077d03231ec72
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/118146
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>