XLSX export: improve handling of checkbox (form controls)
This builds on top of commit 8bf8188a461ca834ca817395ce62449b345a24dc
(tdf#106181 XLSX export: output form controls, 2020-05-13) and adds the
missing VML version which seems to be mandated by Excel 2019.

It is not perfect (e.g. there is still an unwanted border around the
checkbox), but the checkbox has a correct position and its label is
readable, while it was just lost previously.

Change-Id: I08198d068a0eb85061d138719cfc60d73c46398e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/118168
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins