tdf#123569 correct position of rotated Writer image
Error was, that the position pFrameFormat->GetHoriOrient().GetPos()
was treated as left/top of snap rectangle and the associated left/top
of a logic rectangle was calculated from it based on rotation angle.
But actually it is already left/top of logic rectangle.
Nevertheless effectExtent has to be calculated for the rotation.

The fix is based on commit 61a6be3aa3b470e5da3ca7ad4b8d1a6155a6a07f.
A direct backport to LO7.2 is not possible.

Change-Id: I8761b673d1358bb476f1678fd2a104eaec6dcb2d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/118243
Tested-by: Jenkins
Reviewed-by: Regina Henschel <rb.henschel@t-online.de>