doc import: halt earlier for non-paragraph styles
related to work being done for tdf#104239.

This should practically be an NonFunctionalChange
that just avoids some unnecessary processing.

Unit tests prove that these semi-defined styles
are still necessary for inheritance etc.,
but they shouldn't need to be used for selecting
the outline numbering.

The "continue" that I am moving was first added with
commit 02cd405c26a0f110e9a042f537da02acfb720f65
Author: Caolán McNamara on Wed Aug 27 09:19:30 2014 +0100
    OutlineStyles can only be Paragraph Styles

which probably just failed some automated tests,
so that suggests there is plenty of room for improvement here.

Change-Id: I5d4d6ba374f0f271b5248fef3d2d8c0ce62082cb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/118342
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>